# Don't convert Symmetric to full for expm()